Split 'canvas.go' along major data structure boundaries
